// src/docs/schemas/parts/create-part/createPartResponse.shema.ts
const createPartResponseSchema = {
    type: "object",

    properties: {
        id: {
            type: "string",
            description: "Part ID",
            example: "77b1bb18-0252-4bc3-86eb-621019ada645"
        },
        name: {
            type: "string",
            description: "Name of the part.",
            example: "Turbo MK400"
        },

        type: {
            type: "string",
            description: "Part type.",
            enum: [
                "PISTON",
                "CRANKSHAFT",
                "TURBO",
                "RADIATOR",
                "CAMSHAFT",
                "ECU",
                "OIL_SYSTEM"
            ],
            example: "TURBO"
        },

        rarity: {
            type: "string",
            description: "Part rarity.",
            enum: [
                "COMMON",
                "RARE",
                "EPIC",
                "LEGENDARY"
            ],
            example: "EPIC"
        },

        rpm: {
            type: "number",
            description: "Maximum engine RPM generated by this part. Must be between 0 and 10000.",
            minimum: 0,
            maximum: 10000,
            example: 9000
        },

        torque: {
            type: "number",
            description: "Maximum torque generated by this part. Must be between 0 and 1000.",
            minimum: 0,
            maximum: 1000,
            example: 400
        },

        heat: {
            type: "number",
            description: "Heat generated by this part. Must be between 0 and 500.",
            minimum: 0,
            maximum: 500,
            example: 120
        },

        cooling: {
            type: "number",
            description: "Cooling power provided by this part. Must be between 0 and 500.",
            minimum: 0,
            maximum: 500,
            example: 200
        },

        stability: {
            type: "number",
            description: "Engine stability modifier provided by this part. Must be between -100 and 300.",
            minimum: -100,
            maximum: 300,
            example: 50
        },

        durability: {
            type: "number",
            description: "Part durability. Must be between 0 and 100.",
            minimum: 0,
            maximum: 100,
            example: 85
        },

        weight: {
            type: "number",
            description: "Part weight in kilograms. Must be between 0.1 and 100.",
            minimum: 0.1,
            maximum: 100,
            example: 15
        },

        heatRes: {
            type: "number",
            description: "Heat resistance provided by this part. Must be between 0 and 100.",
            minimum: 0,
            maximum: 100,
            example: 75
        },
        createdAt: {
            type: "string",
            format: "date-time",
            description: "Date and time when the engine was created.",
            example: "2026-06-09T00:57:04.699Z"
        }
    },

    required: [
        "id",
        "name",
        "type",
        "rarity",
        "rpm",
        "torque",
        "heat",
        "cooling",
        "stability",
        "durability",
        "weight",
        "heatRes",
        "createdAt"
    ]
}

export default createPartResponseSchema;